+++ MaquetaParaFinal/Clases/Agregar/AgregarPersonal.cs
@@ -1,0 +1,55 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Shapes;
+
+namespace MaquetaParaFinal.View.Agregar
+{
+    public partial class AgregarPersonal : Window
+    {
+
+        private void SoloNumero(object sender, TextChangedEventArgs e)
+        {
+
+        }
+
+        private void ControlarNombre(object sender, TextChangedEventArgs e)
+        {
+
+        }
+
+        private void btnAceptarAgPaciente_Click(object sender, RoutedEventArgs e)
+        {
+
+        }
+
+        private void btnCancelarAgPaciente_Click(object sender, RoutedEventArgs e)
+        {
+
+        }
+
+        private void btnAgregarEspecialidad_Click(object sender, RoutedEventArgs e)
+        {
+
+        }
+        private void btnAgregarCategoria_Click(object sender, RoutedEventArgs e)
+        {
+
+        }
+
+        private void Principal_MouseLeftButtonDown(object sender, MouseButtonEventArgs e)
+        {
+
+        }
+
+    }
+}
