+++ MaquetaParaFinal/View/Agregar/AgregarEspecialidad.xaml
@@ -1,0 +1,32 @@
+﻿<Window x:Class="MaquetaParaFinal.View.Agregar.AgregarEspecialidad"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:MaquetaParaFinal.View.Agregar"
+        mc:Ignorable="d"
+        WindowStyle="None"
+        AllowsTransparency="True"
+        Background="Transparent"
+        WindowStartupLocation="CenterScreen"
+        Title="AgregarEspecialidad" Height="200" Width="400">
+    <Grid x:Name="Principal" MouseLeftButtonDown="Principal_MouseLeftButtonDown" >
+        <Image x:Name="Logo" Width="40" Height="40" HorizontalAlignment="Left" Margin="-7,-7,0,0"  VerticalAlignment="Top" Source="/View/Agregar/logo.png"/>
+        <Border Margin="5,5,0,0" CornerRadius="45" Background="#4F709C" BorderBrush="#8fc643" BorderThickness="5" Grid.ColumnSpan="2">
+            <StackPanel>
+                <Label Content="Agregar Especialidad" FontSize="25" FontFamily="Roboto" FontWeight="Medium" Foreground="#F0F0F0" Margin="20,5,0,0"/>
+                <DockPanel Margin="10,20,10,0">
+                    <Label Content="Nombre*" Margin="0,3,0,0" Width="100" FontWeight="Medium" FontFamily="Roboto" Foreground="#F0F0F0"/>
+                    <Border Margin="0,6,10,0"  HorizontalAlignment="Stretch" Background="#213555" CornerRadius="9">
+                        <TextBox x:Name="txtNombreEspecialidad" Margin="5,0,0,0" TextWrapping="Wrap" Text="Nombre" VerticalAlignment="Top" Background="Transparent" Foreground="#F0F0F0" BorderBrush="Transparent" LostFocus="txtNombreEspecialidad_LostFocus" GotFocus="txtNombreEspecialidad_GotFocus" TextChanged="ControlarNombre" />
+                    </Border>
+                </DockPanel>
+                <DockPanel Width="250" Height="35" Margin="0,25,0,5">
+                    <Button x:Name="btnAceptarAgEspecialidad" Content="Aceptar" Width="140" HorizontalAlignment="Left" Height="35" Style="{StaticResource BotonPrimario}" Click="btnAceptarAgEspecialidad_Click"/>
+                    <Button x:Name="btnCancelarAgEspecialidad" Content="Cancelar" Width="90" HorizontalAlignment="Right" Height="35" Style="{StaticResource BotonSalir}" Click="btnCancelarAgEspecialidad_Click"/>
+                </DockPanel>
+
+            </StackPanel>
+        </Border>
+    </Grid>
+</Window>
